+++ package.json
@@ -1,45 +1,46 @@
 {
-  "name": "@overmindtech/sdp",
-  "version": "0.14.0",
+  "name": "@overmindtech/sdp-js",
+  "version": "0.15.0",
   "description": "Javascript and Typescript libraries for the State Description Protocol",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "types": "./lib/cjs/types/index.d.ts",
+  "main": "./lib/cjs/index.js",
   "scripts": {
+    "clean": "rm -rf ./lib",
+    "build": "npm run clean && npm run build:esm && npm run build:cjs",
+    "build:esm": "tsc -p ./tsconfig.esm.json && mv lib/esm/index.js lib/esm/index.mjs",
+    "build:cjs": "tsc -p ./tsconfig.cjs.json",
+    "prepack": "npm run build",
     "test": "jest"
+  },
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./lib/esm/types/index.d.ts",
+        "default": "./lib/esm/index.mjs"
+      },
+      "require": {
+        "types": "./lib/cjs/types/index.d.ts",
+        "default": "./lib/cjs/index.js"
+      }
+    }
   },
   "keywords": [],
   "author": "Dylan Ratcliffe",
   "license": "ISC",
   "dependencies": {
     "google-protobuf": "^3.21.2",
-    "node": "^19.2.0",
-    "rxjs": "^7.6.0",
     "sha1": "^1.1.1",
     "to-data-view": "^1.1.0",
-    "ts-protoc-gen": "^0.15.0",
-    "uuid": "^9.0.0",
-    "ws": "^8.11.0"
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.5",
-    "@babel/preset-env": "^7.20.2",
-    "@babel/preset-typescript": "^7.18.6",
-    "@types/chai": "^4.3.4",
     "@types/google-protobuf": "^3.15.6",
-    "@types/mocha": "^10.0.1",
+    "@types/jest": "^29.2.4",
     "@types/sha1": "^1.1.3",
     "@types/uuid": "^9.0.0",
-    "@types/ws": "^8.5.3",
-    "babel-jest": "^29.3.1",
-    "chai": "^4.3.7",
-    "create-index": "^2.6.0",
     "jest": "^29.3.1",
-    "jest-environment-jsdom": "^29.3.1",
     "jest-websocket-mock": "^2.4.0",
-    "mocha": "^10.1.0",
-    "nyc": "^15.1.0",
     "ts-jest": "^29.0.3",
-    "ts-node": "^10.9.1",
     "typescript": "^4.9.4"
   },
   "repository": {
